refactor: split up lib.rs
